Rename `hub.py` to `model.py` and add unit tests.
