+++ mmcls/apis/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .hub import ModelHub, get_model, init_model, list_models
 from .inference import inference_model
+from .model import ModelHub, get_model, init_model, list_models
 
 __all__ = [
     'init_model', 'inference_model', 'list_models', 'get_model', 'ModelHub'
